Added Ext.ux.swfbtn (http://www.sencha.com/forum/showthread.php?87910-Ext.ux.swfbtn-SwfUpload-2.2.0-ext-3.0-Flash-10-features)
This is actually shitty code from a scoping and extendability
perspective, but it happens to already have a facility to post a
description with uploaded files, so we're using it for now. I had to
modify its source code to get it to work properly. (I modified swfbtn.js
itself)