Added Ext.ux.RapidApp.WinFormPost + save search updates
Creates new all-purpose JS function Ext.ux.RapidApp.WinFormPost that
is simular to Ext.ux.EditRecordFieldSet but it is used to POST the
values of a fieldset to a URL

Also updated the saved search functionality to use this. Now there is a
checkbox that allows setting to save the search as a new search, or save
the existing search.