Added new Hashref Attribute "no_json_ref_types" that is now checked on calling render_data in the RapidApp Controller class. If the data to render is a reference and its type is in this hash, it will not render as json even if render_as_json is true.
By default, this reference type is in the list:

  IO::File

The reason is that Catalyst is able to provide it's own custom rendering
to serve up these types of objects as file downloads - I can't think of
any time we would want to render an IO::File object as Json, so it is
now excluded by default. If we did for some strange reason need to
render as json, we would just override the no_json_ref_types attribute
when subclassing any desendant of a class with
RapidApp::Role::Controller

This change was made to fix a bug in GreenSheet. I was setting
render_as_json to false in the Attachment file_download action, but this
was getting persisted causing subsequent action requests taht should be
rendered as json to not be. This problem continued even after I set
"no_persist" to true. So this is a separate problem. The change in this
commit is really only a workaround to get around this problem, however,
I think that the "no_json_ref_types" approach is sound and actually a
better way to do this than to toggle the value of "render_as_json". That
doesn't mean the no_persist problem shouldn't also be investiagted and
fixed at some point...