Added code to handle cases where a user session is deleted on the server side, but the client doesn't know. This can happen if a session times out, or if there are multiple windows open and Logout is clicked on one of them.
The new code in this commit that handles this is both in the backend and
the front end. In the back end, the AuthRequire Moose Role now sets the
new HTTP header 'X-RapidApp-Authenticated' to '1' if the user is
authenticated, and '0' if the user is not.

The front end code adds a global event listener ('requestcomplete') to the
Ext.Ajax singleton which checks for this new header and sets it in the
global javascript variable 'Ext.ux.RapidApp.Authenticated' which is
persisted in the running ExtJS/javascript client interface. If the value
of this global variable changes from 1 to 0, then it indicates that the
session has been deleted on the server side, and the javascript function
"window.location.reload();" is executed to reload the page (which always
gets the front end interface in sync with the back end)