Enhanced and extended front-end detection of back-end invalid session defined in the previous commit (see commit ba5b3966bf36baab9538fdcd71fda1d6d737d88d for more details).
Now, instead of simply reloading the page, a "re-login" dialog is
presented and the user is given a chance to re-enter their password and
get logged back in to the system without reloading the javascript
interface (and losing their application state).

Now, instead of storing 0 or 1, the 'X-RapidApp-Authenticated' header
returned by Modules with the AuthRequire Moose role return 0 or the
username, which is stored in the js variable
Ext.ux.RapidApp.Authenticated in the front end to remember the value of
the current username and then used if/when the re-login dialog is
displayed. By design, the relogin dialog does not permit the user to
enter a different username.

One current limimitation is that, depending on the interface element being
clicked on by the user when the system detects the session on the server
is invalid, the interface will not be returned to a perfectly consistent
state even after they have re-authenticated. For example, if the
operation in question is one where a new tab is opened, the new tab will
be blank and will stay blank after they re-login (but subsequent
operations will work normally). This is because currently there is no
way to pause and "repeat" the last operation during this re-login
process. This may be looked into later, but will be very hard to fix as
the way AppTreeExplorer handles tabs is that the new Tab is created by
the front-end before a request to the server is made, so the Tab already
exists by the time the front-end discovers it can't be populated because
the user is not logged in.

ANOTHER MAJOR LIMITATION IN THIS VERSION is that it requires the login
url to be hard-coded in 'Ext.ux.RapidApp.loginUrl' (which defaults to
'/main/banner/auth/login'). This will be fixed by refactoring the main
login dialog code itself, which will include redefining RapidApp
configuration options for a more "sane" auth layout than it has
currently, and the new "relogin" code (which uses
Ext.ux.RapidApp.WinFormPost) will likely be merged with the normal
"login" window code.