New mechanism to handle detecting session timeout
Replaced the recently implemented mechanism that used the
'requestcomplete' callback to handle asking the user to re-login with a
much more robust technique that involves overriding the 'handleResponse'
function to actually intercept the request, have the user log back in,
then re-call the original request so that the UI element loads whatever
as if nothing even happened. This fixes the previous limitation with
this like tabs loading with blank content.