RapidApp errors no longer display js browser error
Still throw a js error, but override window.onerror first to prevent it
from showing up as an error in the browser. This is only reset
temporarily so that other js errors will still show as normal later on.

The reason we throw the js error (which we are basically hiding) is to
halt execution of js code in place so that we don't need to worry about
the data that is getting passed back to whatever js component generated
the request that caused the server side exception.