Redefined parts of the exception handling again.
No longer using throw, and the error dialog is now displayed by using
the Ajax events. Now return status "542" from the server side upon an
exception. This is a better approach because it allows failure handlers to
be called if they exist (where throw would block those errors), and now
that we return a 5xx code, it is seen as an actual failure by the js
code