AppGrid2: "load_saved_search" now called ONREQUEST if exists
If the subclass implements the method "load_saved_search" it will be
called (with no arguments) during "ONREQUEST". This allows subclasses to
hook in, but its up to the subclass to do all the work, such as checking
for params in the request object (i.e. "search_id"), etc.

It is expected that this method will call "batch_apply_opts" to
reconfigure the options before content is rendered