Added X-RapidApp-Warning support
Added another global override to Ext.data.Connection that will check for
the 'X-RapidApp-Warning' response header. The data stored in this header
is expected to be a serialized JSON string of a JavaScript object
containing at least the property "msg" and optionally "title" (if title
is not set it defaults to "Warning"). If this data is present in the
response of any Ajax request, a dialog box will be displayed containing
the serialized title/msg.

This is very similar to "X-RapidApp-Exception" that was previously
added, except the Ajax request is not aborted and is allowed to finish
normally, and the message to display is contained within the
X-RapidApp-Warning directly (while X-RapidApp-Exception is only a flag,
and the message displayed is stored in the normal response body).

X-RapidApp-Warning provides a mechanism to display an arbitrary warning
message, and since it does not interfere with the related request, it is
informational only.

A new method "set_response_warning" has been added to the Controller
Role, which makes it available to all RapidApp modules. This method
accepts an argument (string or hashref representing the JS object) and
handles encoding the JSON and setting the header (including validation).