Expanded 'modules' API to support Hashref or string
Prior to this commit, the "modules" parameter accepted a hash of
key/values, with the key representing the module name, and the value
representing the class name, which could only be a string. While this
API is preserved, the value can now optionally be specified as a
Hashref. In the new latter case, the Hashref should have these 2 keys:

  * class
  * params

The "class" parameter is required, and represents the same class name
string that the old API supports when the value is a string. The params
parameter should be another Hashref, and it will be passed to the "new"
method when the module object is constructed (i.e. class->new(params) )