"custom" field configs for multifilter, store_use_xtype for DataStore
The column model now supports "field_cnf" which if present will be used
for the "datafield" in MultiFilters. This is being added primarily to
support custom Combo fields.

Added "store_use_xtype" bool option to DataStore. If set to true, the
JsonStore attribute will return a hashref with the xtype set to
"jsonstore" instead of returning a RapidApp::JSONFunc object of a 'new
Ext.data.JsonStore'. This is more robust in many cases, including when
using an AppCombo as the field_cnf for multifilters. I added this as an
option because I am not certain if changing as a new default would work
in all situations