Added code in Module to init ONREQUEST_called to true
This was needed to prevent the possibility that ONREQUEST might get
called during object BUILD. If this happes, the the object "defaults"
get set early, leading to unexpected results. It is vital that the first
time ONREQUEST is called it is called from the controller on a real
request, *after* all BUILD methods have fully run.