Added DbicAppCombo + dynamically creating combos for related cols
Added logic with the creation of the related columns to create
DbicAppCombo and setting the field_cnf of each to its content. This
makes all these fields be dropdowns with real possible values in the
lists in the multifilter display.

NOTE: This causes the first page load to be even slower. On GreenSheet,
it now takes nearly a full minute to build the modules on the first page
load.