MultiFilters: Now displays friendly column header names (where defined)
Note: This is not done using the displayField and valueField parameters
of Ext.form.Combo. Instead a map and a reverse map is built and the raw
values are substituted back and forth between the column.name and
column.header properties. The reason is the limitations that exist with
Ext.form.Combo and loading default values. It requires hooking into
store load events, etc. It's actually more work than just doing it this
way (and this is actually leaner).