Start on DataStore2
Starting this branch/approach after giving up on branch
"DataStoreAsClass". RapidApp::DataStore2 is this commit is just a copy
of RapidApp::DataStore from the now defunct DataStoreAsClass git branch.
The original plan was to change DataStore across the board and preserve
the existing API. This proved to be too problematic, so instead this new
DataStore2 module will operate in parallel to DataStore (just like
AppGrid has been operating in parallel to AppGrid2).

DataStore2 will have a different API acting as a separate module unlike
the current DataStore Role. I will then convert AppGrid2 and other new
modules to use DataStore2, and leave the older modules working with the
DataStore Role.