AppCmp: Renamed 'config' to 'extconfig' (but preserved old API)
Changed the name of the 'config' attribute to 'extconfig' to prevent
possible method name conflicts (such as if we merge with the namespace
of a Catalyst Component).

Created mapping subs for all the old names (config and all its native
trait handlers such as apply_config, get_config_param, etc) so no API
change is required yet. Currently either config or extconfig names work.
(later on we will remove the old names which will force the API change
at that time)