Added "ONREQUEST_calls" system to Module
Added a new set of attributes to facilitate easier calling of code in a
predictable order during "ONREQUEST":

 * ONREQUEST_calls
 * ONREQUEST_calls_early
 * ONREQUEST_calls_late

These attributes have Array native traits and setup standard accessor
methods to add and retrieve from them.

These consist of an ordered list of functions (as RapidApp::Handler
objects) that will be called during various stages of ONREQUEST. This
accomplishes the same thing as the previous method to add hooks to
ONREQUEST with method modifiers, except that way could lead to a
confusing order of execution with multiple levels of inheritance.
This new way makes it easy to know what order functions will be called.
If you add the ONREQUEST calls during BUILD of each class, then they
will end up being called in the same order as BUILD, which is intuitive
(parents to subclasses).

The add methods (such as "add_ONREQUEST_calls") accept either a
RapidApp::Handler object, or a string representing the name of the
method in the current object to call.