New JS listener approach - AppCmp: "add_listener"
I discovered that the approach of the "event_handlers" attribute in
AppCmp would not work in all cases, most notably in the case of
Ext.data.Store objects because it does not inherit from Ext.Component,
and only classes that inherit from Ext.Component can have ExtJS plugins.
The plugin feature is relied on by the event_handlers technique, so
because it won't work for stores, I had to go back to the drawing board.

The reason event_handlers was needed at all was because the "listeners"
config attribute is a hash, which meant that multiple handlers for the
same event could not be defined at once.

In this commit, to overcome this limitation, I have added new method
"add_listener" to AppCmp that pushes the handler (RapidApp::JSONFunc
object) into a list under a new config param "listener_callbacks" and
then sets the listener for the given event to a special js function that
iterates over the list for the given event under listener_callbacks and
calls each in order.

Because this is constructing a special config, "apply_listeners" should
no longer be called at all (because it could override the listeners
defined by add_listener). So, going forward, "add_listener" should be
used instead.