Added rudimentary Devel::NYTProf Profiling support
Created new Catalyst module "RapidApp::CatalystX::Profile" based on the
CPAN module CatalystX::Profile. The difference from the CPAN module is
that the RapidApp version must be started with
"/profile/start_profiling" and the CPAN version is hard coded to start
profiling right away. This is not useful when we want to do profiling on
individual requests after the App has already started.

Since the original module was so tiny to begin with, I didn't even
bother with trying to extend it, etc.

For usage, see the POD Synopsis in RapidApp::CatalystX::Profile