Added back legacy support for "no_persist"
The "no_persist" attribute is a throwback from before the recent
rapidapp changes with ONREQUEST, preloading, etc, however, there are
still some apps (GreenSheet) that assume all lazy attributes will be
cleared on every request if "no_persist" is true. This is now the case
again. This setting should be avoided if its not needed because it hurts
performance, and is depricated and will probably be removed at some
point.