Got deleting working again in AppGrid2, now using "flags"
Deleting is now enabled if "action_delete_records" sub is defined and
the Module flag "can_delete" is set. Under DbicAppGrid2, the sub
"delete_rows" should be defined.

Note that this "delete" operation is currrently specific to AppGrid2 and
*NOT* DataStore2. This is different from the "destroy_records" action in
DataStore that is based on Ext's store "destroy" which is limited in the
way it works. Also note that destroy has not even been fully completed in
DataStore2.