REFACTORED per request attr clearing (SIGNIFICANT)
Changed the way per request attributes are reset on every request in
Module. Instead of clearing them at the begining of the request, they
are cleared at the end of the request in ModuleDispatcher. Defaults are
still set on the first call to ONREQUEST, however, instead of doing the
reset at that time, a reference to the module is added to the Catalust
stash. At the end of the request after all Module processing has
completed, this Module list is enumerated and the attributes are cleared
then.

This change makes coding much more straightforward and understandable,
especially when dealing with Modules that manipulate eachother (such as
with DataStore2)

This change is significant, however, it should not change the API.
Testing will be required, however.