AppStoreForm2: support for auto-loading tab on create
Added new feature to AppStoreForm2 that will automatically try to load a
"loadCfg" in a tab if it exists in the response JSON packet on store
create. The idea is that this loadCfg should contain a valid loadTarget
(from AppTab) and will usually contain a loadCfg to open the item that
was just created. This loadCfg currently has to be built in the response
in "create_records" manually. Something like this:

	return {
		success => 1,
		msg => 'Created contact',
		loadCfg => {
			title 	=> $Contact->last . ', ' . $Contact->first,
			iconCls => 'icon-vcard',
			autoLoad => {
				url 	=> $self->parent_by_name('contacts')->Module('item')->base_url,
				params	=> {
					orig_params => $self->json->encode({
						id => $Contact->id
					})
				}
			}

		}
	};

Also, the option 'closetab_on_create' has been switched from being a
Moose Class attribute to an option in the Ext object (defaults to true).

So to set it, call apply_extconfig