Extended AppCombo2 to auto populate the displayValue on load
This code was generalized from the HOPS Workspace combo and is now
available to all AppCombo2 based combos. JS code checks for "value" and
"focusClass" on load. css contained in focusClass is applied if found.
This allows loading a remote combo and having it display the correct
value (even though the underlying data value is different) and also
setting the icon/style.