Made AppCombo2 its own Ext class
This was done for the purpose of supporting applying optional css to the
field *only* when a value is set. Setting "focusClass" causes the css to
be applied even when the field is still empty.

Added new option "value_addClass" which is a string of css classes that
will be applied to the field when "setValue" is called. This is the same
as focusClass, but it doesn't do the other things that Ext will do with
focusClass (notably setting the class on focus). We want to be able to
se the class on setValue, so I had to break down and write an extended
class to hook into this.