Added 'CustomPrompt' feature
This feature utilizes the same technique as the "ReAuth" feature that
allows the system to allow a user to log back in if their session is
expired, and still complete the original request...

The CustomPrompt feature is more open ended; on every Ajax request the
'X-RapidApp-CustomPrompt' is checked for in the response, and if it is
there, a custom form in a window is displayed based on the serialized
data contained therein.

This defines a simple API that allow specifying an arbitrary set of form
fields and buttons. When the user clicks any of the buttons (except
cancel) the data from the form is submitted back to the server along
with a replay of the original request. (the CustomPrompt data is by
default stored as JSON in a request param named 'customprompt').

On the backend, there new exception class RapidApp::Error::CustomPrompt
has been defined which builds the data to be sent back to the client in
the header.

To use this feature, simply call die passing a
RapidApp::Error::CustomPrompt object. The idea is you can check for the
presence of the "customprompt" param, and if its not there, or doesn't
contain what you want, you can reprompt the user by calling die. Then
since the original request is replayed, you can pick back up on the
backend code where you left off when the request comes back.