Now CustomPrompt data is sent back in HTTP headers
Instead of appending the "customprompt" param to the HTTP request, the
value of the button and fields are now stored in the HTTP Request
Headers 'X-RapidApp-CustomPrompt-Button' and
'X-RapidApp-CustomPrompt-Data' respectively.

Additionally, new attributes have been added to the Module Role to access
this data: $self->customprompt_button and $self->customprompt_data. The
customprompt_data attribute is a HashRef and represents the deserialized
field values entered by the user. So, for example, if a CustomPrompt
were sent back with a textfield named "foo" once the user enters it and
the request comes back in, this will be available via accessing
$self->customprompt_data->{foo} which makes it just as easy to get to as
$self->c->req->params->{foo} (in the case of foo being sent as a query
param)