Added X-RapidApp-Callback feature
On all Ajax requests the new header X-RapidApp-Callback is checked for
in the response from the server. This should conatin a JSON encoded
object with these possible properties:

  * func
  * anonfunc
  * arguments

Either func or anonfunc are required. func should be a string
representing an existing javascript function name to be called, anonfunc
is an anonymous function that is evaled (i.e. a RapidApp::JSONFunc
object on the server side before serialization). arguments, optional,
should be an array and if present it will be passed to the function as
its argument list.

If present the function in X-RapidApp-Callback will be called upon
successful Ajax request in addition to anything else that is called at
this time. Unlike features such as X-RapidApp-CustomPrompt, the original
request is not interupted or repeated.

This just provides a simple way to have any function called at the end
of a request. The function is called with the scope (this reference) of
the Ext.data.Connection object that generated the request. Additionally,
this.response is set to the response object.

The JS side is complete in this commit. Next I will add a convenience
method to the backend perl code...