Refactored and completed X-RapidApp-Callback
Had to do some major refactoring of the JS code from the last commit
because setting the scope breaks many functions. Problem is you can't
call Function.apply() without setting the scope, and apply provides the
ability to turn an array into a parameter list (i.e.
func.apply(this,array)). I could not figure out any other way to do thsi
without apply other than manually building an arg list and then calling
eval on it, which is what I ended up doing...

Since the scope feature may still be wanted, this is now controllable by
setting the "scoped" option to true, which makes the total possible
options now:

  * func
  * anonfunc
  * arguments
  * scoped

On the backend, I have added the "set_response_callback" method to
Controller, and this has a very nice API, here are a few examples:

-----
$self->set_response_callback( 'Ext.ux.MyFunc' );

$self->set_response_callback( alert => 'foo!' );

$self->set_response_callback(
	'Ext.Msg.alert' => ( 'A message!!', 'this is awesome!!' )
);

my $func = RapidApp::JSONFunc->new(
	raw => 1, func => 'function(){ console.log("anon!!"); console.dir(this.response); }'
);
$self->response_callback_scoped(1);
$self->set_response_callback(
	$func => ( "arg1",{ key_in_arg2 =>'blah!!!' },'arg3',\1  )
);
-----

The set_response_callback method handles setting up the data in the
X-RapidApp-Callback header. The first argument is the function and the
remaining arguments are the arguments to the function. If the function
is a RapidApp::JSONFunc object it is set as the anonfunc instead of
func. To use the "scoped" option, set the new attribute
response_callback_scoped to true (defaults to false).

Note that when using a RapidApp::JSONFunc object as the func, it must be
"raw" (that is the only thing that is useful because if you are calling
an existing function, you would just pass its name as a string in the
first arg