DbicLink/DbicExtQuery new feature: 'read_extra_search_set'
DbicLink now checks for the existance of read_extra_search_set method,
and if it exists, it calls it and passes it as the second argument to
the "data_fetch" method. In DbicExtQuery, if this new second arg exists
it is added verbatim to the filter_search set.

This allows defining read_extra_search_set in consuming classes as
another way to control the search on the fly, without needing to pass
params in from the Ajax client.

The code in DbicExtQuery has gotten pretty messy, and it should be
refactored at some point