listener_callbacks now fully working, eliminated "event_handlers"
The "listener_callbacks" feature replaced "event_handlers" and was
needed becase event_handlers used plugins and didn't work with objects
that didn't inherit from Component (such as the Store). The
listener_callbacks works with all Observable objects, including
stores...

However, it wasn't fully working - it didn't work with the click event
in TreePanel, and so event_handlers was left in place. The problem with
the Tree click event was that the first argument of that event is the
"node" and not the component, which was assumed by the previous
listener_callbacks logic. This was basically just a scope problem; added
logic to check for and accomidate this, and now it works in all cases...

With this, the event_handlers feature has been totally removed and the
remaining modules that used it have been convered to use
listener_callbacks (with the "add_listener" method)