REFACTORED RapidApp::DBIC::ResultSet::BaseConditions
BaseConditions now uses a chained ResultSet approach in the search_rs
override to get the ResultSet class itself to do the search and attr
merging. This fixes the bug with the previous implementation that would
break when trying to add other attributes such as group_by and when
"count" is called.

Also changed the API; two main attributes can be defined:

  * base_search_conditions
  * base_search_attrs

These are used to create an intermediate ResultSet that will be chained
to the *first* call of the original search_rs. To prevent subsequent
chained ResultSets to be chained with the base_search again, the key
"BaseConditions" is stuffed into the attrs which get passed on to
subsequent chained ResultSets. The object attrs are then inspected and
if this key is found, the base_search ResultSet is not chained.

Also expanded DbicExtQuery and DbicLink to support a 'get_ResultSet'
attribute that can be used to to get a custom ResultSet on every
data_fetch. This didn't end up being needed because of the new
BaseConditions design, but since its a useful feature I left it in.