Added optional 'literal_dbf_colname' option to DbicExtQuery
By default DbicExtQuery prepends 'me.' to column names that are not
already associated with a joined table (presumed to be columns of the
main selected table). This causes problems in special cases, such as
with "virtual" columns that are actually function calls.

Column names specified in literal_dbf_colname will not have me.
prepended...