MultiFilters: New special condition "is"
New special condition "is" which is treated the same as "is equal to"
except when "is" is selected the data field is transformed into a
rel_combo dropdown if the "rel_combo_field_cnf" property is set on the
column. (previously just "field_cnf")

This changes the behavior where the rel combo would be set always if it
existed. This made it impossible to search for partial strings (with
contains, starts with, etc). It was only possible to use the dropdown.
With the new "is" condition, either can be used. This makes MultiFilters
much more useful in general