dbf_virtual_fields: added ability to specify 'composit_fields'
Extended the dbf_virtual_fields feature to allow defining the
"composit_fields" that are associated with each dbf_virtual_field. The
composit fields should be the columns that are used in the raw SQL
function defined. They must be available to teh function, so they must
be selected along with the virtual field.

This feature was added for performance. Previously, these composit
fields had to be added to always_fetch_columns, which meant those
columns (and their related joins) would *always* be fetched. With this
new feature, they will be fetched automatically with the virtual field,
or not fetched if the virtual field is not fetched.

The new API to define dbf_virtual_fields in a DbicAppGrid2/DbicLink
class looks like this:

---------------
has '+dbf_virtual_fields' => ( default => sub {{
	contacts_list_concat => {
		function => \[
			# Literal SQL:
			'GROUP_CONCAT(' .
				'DISTINCT CONCAT(contactlnk.first," ",contactlnk.last) ' .
					'order by contactlnk.last ASC SEPARATOR ", "' .
			') AS contacts_list_concat'
		],
		composit_fields => [
			'contactlnk_first',
			'contactlnk_last'
		]
	}
}});
---------------

The old API still works and can be used if no composit fields are needed:

---------------
has '+dbf_virtual_fields' => ( default => sub {{
	contacts_list_concat => \[
			# Literal SQL:
			'GROUP_CONCAT(' .
				'DISTINCT CONCAT(contactlnk.first," ",contactlnk.last) ' .
					'order by contactlnk.last ASC SEPARATOR ", "' .
			') AS contacts_list_concat'
		]
}});
---------------