Added Ext.ux.RapidApp.Plugin.HtmlEditor.LoadHtmlFile
This HtmlEditor plugin allows you to load the editor content with an
Html file from your computer.

It uses the new server side SimpleCAS action "upload_echo_base64" which
takes a file upload, converts it to base64, and sends it back to the
client within a json packet.

Currently does not handle unicode