Added Ext.ux.RapidApp.Plugin.ClickableLinks - refs #1
Added new general plugin that adds a click listener to get normal links
working in Ext components. Works in DataViews, DOES NOT WORK IN
HtmlEditor yet.

Also expanded SimpleCAS "fetch_content" method to allow an additional
argument on the URL after the checksum that if supplied will be treated
as the filename and the Content-Disposition will be set to "attachment"
instead of "inline". This prevents the browser from navigating to the
link target.