REFACTORED AND ENHANCED AUTOPANEL - (Ext.ux.AutoPanel)
Refactored and cleaned up the Ext.ux.AutoPanel class for the first time
since it was written when RapidApp was still very young (and I didn't
know what I was doing yet in ExtJS).

Ext.ux.AutoPanel is used everywhere and is one of the most important JS
classes in all of RapidApp. It still does the same basic thing as before
but isn't as ugly and is probably a bit cheaper now as well.

Added a new optional 'cmpConfig' param that can be passed in and will be
merged (taking precidence) with the component config returned by the
server. This allows doing things like setting "listeners" and being able
to apply them to the dynamically loaded component.