AutoPanel: Added optional "cmpListeners" config param
More enhancements to AutoPanel (see previous commit). Added config
parameter 'cmpListeners' that if defined will be *added* to the
listeners of the dynamically loaded component when "initComponent" is
called. This is superior to setting "listeners" in cmpConfig because,
unlike that technique, cmpListeners won't overwrite the existing
listeners that may exist in the config sent by the server.