FIXED SPORADIC 'store is undefined' JS exception (I *think*)
Added a check to make sure the store is found before calling load on it.
There are certain edge cases where the store isn't found by lookup by
the storeId. This is probably related to multiple requests and changing
auto generated store Ids sent from the server side. This is a deeper
problem, but it appears that the only times this happens is during edge
cases where it doesn't cause any harm to not call load, but the
exception does cause problems.

Now the exception shouldn't happen, but it will silently not load() on
the store. This needs to be revisited (I placed a comment in the code
about this)