Refactored AppCombo2 js from online code snippet
Refactored AppCombo2 a bit using the code snippet posted here:

 http://www.sencha.com/forum/showthread.php?128164-Set-value-on-a-searching-combo-box-SOLVED&highlight=combo+query+type

This is basically the same technique I was already using, but it has
some specific if/else logic and checks and solves a couple of the issues
that AppCombo2 was having. I integrated AppCombo2 specific logic into
the code example and its working really well now for searching and
selecting.