Added a number of CSS fixed for Web 1.0 rendering - refs #43 of hops
Right-justified field labels.  (this ought to be based on config, but no time for that)
Made checkbox into a xt-textfield rather than its own class.
This fixes its alignment to be like everything else.
Added a .reset class to make the content of an htmleditor more likely to render correctly