HTMLSnippet: Added support for "wrapping" for style isolation
New option "wrap_snippet" if true will generate a class name and wrap
the html with it while also adding that class name to every css
selector. This is an alternative to direct inlining that allows use of
css classes but prevents them from conflicting with existing/other
defined classes and styles.