Added methods to SimpleCAS/Store to add/fetch from plain scalars rather than files - refs #69
SimpleCAS is geared toward reading and writing files, however, when we make the next store
which goes to the DB, it will be a waste to read/write form the filesystem for some operations.
I added add_content and some helper methods so that we can store a file's contents directly.