Added 'max_pagesize' option to DataStore and set AppGrid2 to use
AppGrid2 now sets ext pageSize *and* DataStore's max_pagesize to 25.
DataStore will adjust the store read params for "limit" and "start" if
they are not supplied or if limit is greater than max_pagesize. This
prevents the front end from being able to request a page of data greater
than max_pagesize.

Ext store.load in the grid currently makes a request with no
start/limit. This commit stops this operation from fetching ALL records
which can be quite slow depending on the total number.