Added 'X-RapidApp-ServerEvents' header + controller method
Added a new special RapidApp response header for server events (see
previous commits). This header is set with the new Controller method
'add_response_server_events' which can be called at any time and passed
a list of event names. These event names will be sent back as a JSON
Array in the X-RapidApp-ServerEvents header and then fired on the client
side.