Now recording request path and parameters, and current user in ErrorReports
I also fixed a problem where the timestamp of the report wasn't generated until after it got serialized,
causing all error reports to display with 'now' as the timestamp.

Fixed timezone display in the report view, but not in the AppGrid.  Will fix this more completely when
we add proper timezone support to RapidApp.