Added new sugar functions "userexception" and "rawhtml" - refs HOPS #131
Added RawHtml to wrap a string with a simple class that indicates it is html, and stringifies back to the html scalar.

Improved the usererr function to handle html messages and titles.

Added userexception to mimic usererr, but log the exceptions as real errors.

Removed a bunch of cruft and dead code from RapidApp::Error

Removed WrappedError altogether.

Improved the way View::JSON and View::HttpStatus work.

Added Web1 form for submitting user comments to an exception.

Fixed timezone support for client-rendered timezones, fixing ErrorBrowse in the process.