Made trace capture file handling much more robust - refs GreenSheet #42
The problem was that the trace capture was trying to overwrite a file without
checking to see who owned the file.  If the file had been created as root, Apache
would not be able to overwrite it, and would fail on startup.

I added logic to first try deleting the file, then try appending the PID to the
filename if the original couldn't be deleted.