Added support for 'renderer' and 'extra_meta_data' to RapidApp::Column
renderer can now be specified instead of render_fn. renderer has to be a
JS function (i.e. with RapidApp::JSONFunc) but render_fn can still be a
string of a function name

Also 'extra_meta_data' can be specified and can be anything. Its just an
optional data holder.