Fixed bug in 'rawhtml' usererr + modified errMsgHandler
There was a type constraint of 'Str' on the userMessage attribute of the
UserError object that would have prevented the rawhtml function from
ever working. I changed this to 'Str|Object' and now this is no longer
throwing an exception.

I also enhanced the user error message box to use Ext.Msg.show instead
of Ext.Msg.alert so I could pass some additional options which include
an error icon and a minWidth of 275