DbicLink: id_in now built-in request parameter - IMPORTANT!
A bit of refactoring, no longer calling "data_fetch" method in
DbicExtQuery. Now calling a method to get the ResultSet object and then
building th data packet within DbicLink.

The request parameter 'id_in' is now built-in across the board, much
like 'columns', 'dir' and 'sort'. If id_in is passed, the search will be
restricted only to rows whose 'record_pk' is within the list of values
supplied in id_in.

I don't think this will cause an API change, but it is a significant
change so testing is required.

id_in was previously specific only to DbicAppGridSelector.